// backend/src/generated/interfaces.ts
import { ApiProperty } from "@nestjs/swagger";

export const JsonValueApiProperty = (options: Parameters<typeof ApiProperty>[0] = {}) => {
  return ApiProperty({
    ...options,
    oneOf: [
      { type: 'string' },
      { type: 'number' },
      { type: 'boolean' },
      { 
        type: 'object',
        additionalProperties: true
      },
      { 
        type: 'array',
        items: {
          oneOf: [
            { type: 'string' },
            { type: 'number' },
            { type: 'boolean' },
            { type: 'object' },
            { type: 'array', items: { type: 'object' } }
          ]
        }
      }
    ],
    nullable: true
  });
};

// This file was auto-generated by prisma-generator-typescript-interfaces

export type Role = "student" | "mentor" | "admin";

export class User {
  @ApiProperty({ required: true, type: String })
  id: string;

  @ApiProperty({ required: true, type: String })
  firstName: string;

  @ApiProperty({ required: false, type: String })
  middleName: string | null;

  @ApiProperty({ required: true, type: String })
  lastName: string;

  @ApiProperty({ enumName: 'Role', enum: ['student', 'mentor', 'admin'] })
  role: Role;

  @ApiProperty({ required: true, type: Date })
  createdAt: Date;

  @ApiProperty({ required: true, type: Date })
  updatedAt: Date;

  @ApiProperty({ required: false, type: Date })
  deletedAt: Date | null;
}

export class UserAccount {
  @ApiProperty({ required: true, type: String })
  id: string;

  @ApiProperty({ required: true, type: String })
  userId: string;

  @ApiProperty({ required: true, type: String })
  authUid: string;

  @ApiProperty({ required: false, type: String })
  email: string | null;

  @ApiProperty({ required: true, type: Date })
  createdAt: Date;

  @ApiProperty({ required: true, type: Date })
  updatedAt: Date;

  @ApiProperty({ required: false, type: Date })
  deletedAt: Date | null;
}

export class UserDetails {
  @ApiProperty({ required: true, type: String })
  id: string;

  @ApiProperty({ required: true, type: String })
  userId: string;

  @ApiProperty({ required: false, type: Date })
  dob: Date | null;

  @ApiProperty({ required: false, type: String })
  gender: string | null;

  @ApiProperty({ required: true, type: Date })
  createdAt: Date;

  @ApiProperty({ required: true, type: Date })
  updatedAt: Date;

  @ApiProperty({ required: false, type: Date })
  deletedAt: Date | null;
}